preprocess done

+++ src/s/sc_SpecialComment.h
@@ -1,5 +1,5 @@
 // needs: <vector>, <string>
 struct SpecialComment {
-    int posInPPSource;
+    size_t posInPPSource;
     std::vector<std::string> needs;
 };
